moved the provider tracker to the balance package, to make it reusable on consumer side
