+++ cmd/di.go
@@ -55,7 +55,6 @@ import (
 	"github.com/mysteriumnetwork/node/services/openvpn/discovery/dto"
 	"github.com/mysteriumnetwork/node/session"
 	"github.com/mysteriumnetwork/node/session/balance"
-	balance_provider "github.com/mysteriumnetwork/node/session/balance/provider"
 	session_payment "github.com/mysteriumnetwork/node/session/payment"
 	payment_factory "github.com/mysteriumnetwork/node/session/payment/factory"
 	payments_noop "github.com/mysteriumnetwork/node/session/payment/noop"
@@ -329,7 +328,7 @@ func newSessionManagerFactory(
 			}
 
 			// TODO: the ints and times here need to be passed in as well, or defined as constants
-			tracker := balance_provider.NewBalanceTracker(&timeTracker, amountCalc, 0)
+			tracker := balance.NewBalanceTracker(&timeTracker, amountCalc, 0)
 			validator := validators.NewIssuedPromiseValidator(consumerID, receiverID, issuerID)
 			return session_payment.NewSessionBalance(sender, tracker, promiseChan, time.Second*5, time.Second*1, validator, promiseStorage, consumerID, receiverID, issuerID), nil
 		}
