+++ session/balance/balance_tracker_test.go
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package provider
+package balance
 
 import (
 	"testing"
